+++ packages/patrol_finders/lib/src/custom_finders/patrol_tester.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -334,9 +332,12 @@ class PatrolTester {
     Duration? visibleTimeout,
     Duration? settleTimeout,
   }) {
-    if (!kIsWeb && Platform.isIOS && kReleaseMode) {
+    if (!kIsWeb) {
       // Fix for enterText() not working in release mode on real iOS devices.
       // See https://github.com/flutter/flutter/pull/89703
+      // Also a fix for enterText() not being able to interact with the same
+      // textfield 2 times in the same test.
+      // See https://github.com/flutter/flutter/issues/134604
       tester.testTextInput.register();
     }